archsig の schema 型を分離する
