+++ tools/archsig/src/lib.rs
@@ -4,31 +4,12 @@ use std::ffi::OsStr;
 use std::fs;
 use std::path::{Component as PathComponent, Path, PathBuf};
 
-use serde::{Deserialize, Serialize};
+use serde::Deserialize;
 use serde_json::Value;
 use walkdir::{DirEntry, WalkDir};
 
-pub const SCHEMA_VERSION: &str = "archsig-sig0-v0";
-pub const COMPONENT_KIND: &str = "lean-module";
-pub const VALIDATION_REPORT_SCHEMA_VERSION: &str = "component-universe-validation-report-v0";
-pub const EMPIRICAL_DATASET_SCHEMA_VERSION: &str = "empirical-signature-dataset-v0";
-pub const SIGNATURE_SNAPSHOT_STORE_SCHEMA_VERSION: &str = "signature-snapshot-store-v0";
-pub const SIGNATURE_DIFF_REPORT_SCHEMA_VERSION: &str = "signature-diff-report-v0";
-pub const AIR_SCHEMA_VERSION: &str = "aat-air-v0";
-pub const AIR_VALIDATION_REPORT_SCHEMA_VERSION: &str = "aat-air-validation-report-v0";
-pub const FEATURE_EXTENSION_REPORT_SCHEMA_VERSION: &str = "feature-extension-report-v0";
-pub const THEOREM_PRECONDITION_CHECK_REPORT_SCHEMA_VERSION: &str =
-    "theorem-precondition-check-report-v0";
-pub const RUNTIME_EDGE_EVIDENCE_SCHEMA_VERSION: &str = "runtime-edge-evidence-v0";
-pub const RUNTIME_PROJECTION_RULE_VERSION: &str = "runtime-edge-projection-v0";
-pub const RELATION_COMPLEXITY_CANDIDATE_SCHEMA_VERSION: &str = "relation-complexity-candidates/v0";
-pub const RELATION_COMPLEXITY_OBSERVATION_SCHEMA_VERSION: &str =
-    "relation-complexity-observation/v0";
-pub const RELATION_COMPLEXITY_RULE_SET_VERSION: &str = "relation-complexity-rules/v0";
-pub const EXTRACTOR_NAME: &str = "archsig";
-pub const EXTRACTOR_VERSION: &str = env!("CARGO_PKG_VERSION");
-pub const RULE_SET_VERSION: &str = "sig0-v0";
-pub const DEFAULT_UNIVERSE_MODE: &str = "local-only";
+pub mod schema;
+pub use schema::*;
 
 const DATASET_SIGNATURE_AXES: [&str; 16] = [
     "hasCycle",
@@ -66,1208 +47,6 @@ const DATASET_DELTA_AXES: [&str; 15] = [
     "runtimePropagation",
     "relationComplexity",
 ];
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Sig0Document {
-    pub schema_version: String,
-    pub root: String,
-    pub component_kind: String,
-    pub components: Vec<Component>,
-    pub edges: Vec<Edge>,
-    pub policies: Policies,
-    pub signature: Signature,
-    pub metric_status: BTreeMap<String, MetricStatus>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub policy_violations: Vec<PolicyViolation>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub runtime_edge_evidence: Vec<RuntimeEdgeEvidence>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub runtime_dependency_graph: Option<RuntimeDependencyGraphProjection>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct Component {
-    pub id: String,
-    pub path: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub struct Edge {
-    pub source: String,
-    pub target: String,
-    pub kind: String,
-    pub evidence: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Policies {
-    pub boundary_allowed: Vec<String>,
-    pub abstraction_allowed: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub policy_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub schema_version: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub boundary_group_count: Option<usize>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub abstraction_relation_count: Option<usize>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Signature {
-    pub has_cycle: usize,
-    pub scc_max_size: usize,
-    pub max_depth: usize,
-    pub fanout_risk: usize,
-    pub boundary_violation_count: usize,
-    pub abstraction_violation_count: usize,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct MetricStatus {
-    pub measured: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub reason: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub source: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct PolicyViolation {
-    pub axis: String,
-    pub source: String,
-    pub target: String,
-    pub evidence: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub source_group: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub target_group: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub relation_ids: Option<Vec<String>>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RuntimeEdgeEvidence {
-    pub source: String,
-    pub target: String,
-    pub label: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub failure_mode: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub timeout_budget: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub retry_policy: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub circuit_breaker_coverage: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub confidence: Option<String>,
-    pub evidence_location: RuntimeEvidenceLocation,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RuntimeEvidenceLocation {
-    pub path: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub line: Option<usize>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub symbol: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RuntimeDependencyGraphProjection {
-    pub projection_rule: String,
-    pub edge_kind: String,
-    pub edges: Vec<Edge>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ComponentUniverseValidationReport {
-    pub schema_version: String,
-    pub input: ValidationInput,
-    pub universe_mode: String,
-    pub summary: ValidationSummary,
-    pub checks: Vec<ValidationCheck>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub warnings: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ValidationInput {
-    pub schema_version: String,
-    pub path: String,
-    pub root: String,
-    pub component_kind: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ValidationSummary {
-    pub result: String,
-    pub component_count: usize,
-    pub local_edge_count: usize,
-    pub external_edge_count: usize,
-    pub failed_check_count: usize,
-    pub warning_check_count: usize,
-    pub not_measured_check_count: usize,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ValidationCheck {
-    pub id: String,
-    pub title: String,
-    pub result: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub reason: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub count: Option<usize>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub examples: Vec<ValidationExample>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub metric: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub lean_boundary: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ValidationExample {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub component_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub path: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub source: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub target: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub evidence: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct EmpiricalDatasetInput {
-    pub repository: RepositoryRef,
-    pub pull_request: PullRequestRef,
-    pub pr_metrics: PullRequestMetrics,
-    #[serde(default)]
-    pub issue_incident_links: Vec<IssueIncidentLink>,
-    #[serde(default)]
-    pub analysis_metadata: AnalysisMetadata,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct EmpiricalSignatureDatasetV0 {
-    pub schema_version: String,
-    pub repository: RepositoryRef,
-    pub pull_request: PullRequestRef,
-    pub signature_before: SignatureSnapshot,
-    pub signature_after: SignatureSnapshot,
-    pub delta_signature_signed: NullableSignatureIntVector,
-    pub metric_delta_status: BTreeMap<String, MetricDeltaStatus>,
-    pub pr_metrics: PullRequestMetrics,
-    pub issue_incident_links: Vec<IssueIncidentLink>,
-    pub analysis_metadata: AnalysisMetadata,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RepositoryRef {
-    pub owner: String,
-    pub name: String,
-    pub default_branch: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct PullRequestRef {
-    pub number: usize,
-    pub author: String,
-    pub created_at: String,
-    pub merged_at: Option<String>,
-    pub base_commit: String,
-    pub head_commit: String,
-    pub merge_commit: Option<String>,
-    pub labels: Vec<String>,
-    pub is_bot_generated: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct PullRequestMetrics {
-    pub changed_files: usize,
-    pub changed_lines_added: usize,
-    pub changed_lines_deleted: usize,
-    pub changed_components: Vec<String>,
-    pub review_comment_count: usize,
-    pub review_thread_count: usize,
-    pub review_round_count: usize,
-    pub first_review_latency_hours: Option<f64>,
-    pub approval_latency_hours: Option<f64>,
-    pub merge_latency_hours: Option<f64>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct IssueIncidentLink {
-    pub kind: String,
-    pub id: String,
-    pub url: Option<String>,
-    pub severity: Option<String>,
-    pub labels: Vec<String>,
-    pub opened_at: Option<String>,
-    pub closed_at: Option<String>,
-    pub affected_components: Vec<String>,
-    pub rollback: Option<bool>,
-    pub reopened: Option<bool>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AnalysisMetadata {
-    pub signature_after_commit_role: String,
-    pub excluded_from_primary_analysis: bool,
-    pub exclusion_reasons: Vec<String>,
-    pub primary_cost_record: Option<String>,
-    pub relation_complexity_components: Option<RelationComplexityComponents>,
-    pub runtime_metrics: Option<RuntimeMetrics>,
-    pub alternate_signature_after: Option<Box<SignatureSnapshot>>,
-}
-
-impl Default for AnalysisMetadata {
-    fn default() -> Self {
-        Self {
-            signature_after_commit_role: "head".to_string(),
-            excluded_from_primary_analysis: false,
-            exclusion_reasons: Vec::new(),
-            primary_cost_record: None,
-            relation_complexity_components: None,
-            runtime_metrics: None,
-            alternate_signature_after: None,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RuntimeMetrics {
-    pub runtime_graph_measured: bool,
-    pub runtime_edge_evidence_count: Option<usize>,
-    pub runtime_pair_count: Option<usize>,
-    pub runtime_fanout: Option<usize>,
-    pub unprotected_runtime_propagation_radius: Option<usize>,
-    pub circuit_breaker_coverage_ratio: Option<f64>,
-    pub protected_pair_count: Option<usize>,
-    pub partial_pair_count: Option<usize>,
-    pub unprotected_pair_count: Option<usize>,
-    pub unknown_coverage_pair_count: Option<usize>,
-    pub failure_mode_taxonomy_version: Option<String>,
-    pub coverage_policy_version: Option<String>,
-    pub confidence_threshold: Option<f64>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RelationComplexityComponents {
-    pub constraints: usize,
-    pub compensations: usize,
-    pub projections: usize,
-    pub failure_transitions: usize,
-    pub idempotency_requirements: usize,
-}
-
-impl RelationComplexityComponents {
-    fn relation_complexity(&self) -> usize {
-        self.constraints
-            + self.compensations
-            + self.projections
-            + self.failure_transitions
-            + self.idempotency_requirements
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RelationComplexityCandidateFile {
-    pub schema_version: String,
-    pub repository: String,
-    pub revision: String,
-    pub measurement_universe: RelationComplexityMeasurementUniverse,
-    pub workflow: RelationComplexityWorkflow,
-    #[serde(default)]
-    pub evidence: Vec<RelationComplexityEvidence>,
-    #[serde(default)]
-    pub evidence_candidates: Vec<RelationComplexityEvidence>,
-    #[serde(default)]
-    pub excluded_evidence: Vec<RelationComplexityExcludedEvidence>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RelationComplexityObservation {
-    pub schema_version: String,
-    pub repository: String,
-    pub revision: String,
-    pub measurement_universe: RelationComplexityMeasurementUniverse,
-    pub workflow: RelationComplexityWorkflow,
-    pub counts: RelationComplexityComponents,
-    pub relation_complexity: usize,
-    pub evidence: Vec<RelationComplexityEvidence>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub excluded_evidence: Vec<RelationComplexityExcludedEvidence>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RelationComplexityMeasurementUniverse {
-    pub root: String,
-    pub languages: Vec<String>,
-    pub frameworks: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub rule_set_version: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RelationComplexityWorkflow {
-    pub id: String,
-    pub name: String,
-    pub component: String,
-    pub entrypoints: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RelationComplexityEvidence {
-    pub id: String,
-    pub path: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub symbol: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub line: Option<usize>,
-    pub tags: Vec<String>,
-    pub ownership: String,
-    pub review_status: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub notes: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RelationComplexityExcludedEvidence {
-    pub path: String,
-    pub reason: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub symbol: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub line: Option<usize>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GitCommitRef {
-    pub sha: String,
-    pub role: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SignatureSnapshot {
-    pub commit: GitCommitRef,
-    pub extractor: ExtractorRef,
-    pub signature: ArchitectureSignatureV1DatasetShape,
-    pub metric_status: BTreeMap<String, MetricStatus>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ExtractorRef {
-    pub name: String,
-    pub version: String,
-    pub rule_set_version: String,
-    pub policy_version: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ArchitectureSignatureV1DatasetShape {
-    pub has_cycle: usize,
-    pub scc_max_size: usize,
-    pub max_depth: usize,
-    pub fanout_risk: usize,
-    pub boundary_violation_count: usize,
-    pub abstraction_violation_count: usize,
-    pub scc_excess_size: usize,
-    pub max_fanout: usize,
-    pub reachable_cone_size: usize,
-    pub weighted_scc_risk: Option<usize>,
-    pub projection_soundness_violation: Option<usize>,
-    pub lsp_violation_count: Option<usize>,
-    pub nilpotency_index: Option<usize>,
-    pub runtime_propagation: Option<usize>,
-    pub relation_complexity: Option<usize>,
-    pub empirical_change_cost: Option<usize>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SignatureSnapshotStoreRecordV0 {
-    pub schema_version: String,
-    pub repository: SnapshotRepositoryRef,
-    pub revision: RepositoryRevisionRef,
-    pub scan: ScanMetadata,
-    pub extractor: SnapshotExtractorMetadata,
-    pub policy: SnapshotPolicyMetadata,
-    pub signature: ArchitectureSignatureV1DatasetShape,
-    pub metric_status: BTreeMap<String, MetricStatus>,
-    pub validation_summary: SnapshotValidationSummary,
-    pub artifacts: SnapshotArtifacts,
-    pub analysis_metadata: SnapshotAnalysisMetadata,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SnapshotRepositoryRef {
-    pub owner: String,
-    pub name: String,
-    pub default_branch: String,
-    pub remote_url: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RepositoryRevisionRef {
-    pub sha: String,
-    #[serde(rename = "ref")]
-    pub ref_name: Option<String>,
-    pub branch: Option<String>,
-    pub committed_at: Option<String>,
-    pub parent_shas: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ScanMetadata {
-    pub scanned_at: String,
-    pub scanner_id: Option<String>,
-    pub trigger: String,
-    pub root: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SnapshotExtractorMetadata {
-    pub name: String,
-    pub version: String,
-    pub rule_set_version: String,
-    pub input_schema_version: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SnapshotPolicyMetadata {
-    pub policy_id: Option<String>,
-    pub schema_version: Option<String>,
-    pub version: Option<String>,
-    pub source_path: Option<String>,
-    pub content_hash: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SnapshotValidationSummary {
-    pub schema_version: Option<String>,
-    pub result: String,
-    pub universe_mode: Option<String>,
-    pub failed_check_count: Option<usize>,
-    pub warning_check_count: Option<usize>,
-    pub not_measured_check_count: Option<usize>,
-    pub report_path: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SnapshotArtifacts {
-    pub extractor_output_path: Option<String>,
-    pub validation_report_path: Option<String>,
-    pub policy_path: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SnapshotAnalysisMetadata {
-    pub excluded_from_primary_diff: bool,
-    pub exclusion_reasons: Vec<String>,
-    pub tags: Vec<String>,
-    pub notes: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct SnapshotRecordInput {
-    pub repository: SnapshotRepositoryRef,
-    pub revision: RepositoryRevisionRef,
-    pub scan: ScanMetadata,
-    pub policy_version: Option<String>,
-    pub policy_source_path: Option<String>,
-    pub policy_content_hash: Option<String>,
-    pub extractor_output_path: Option<String>,
-    pub validation_report_path: Option<String>,
-    pub tags: Vec<String>,
-    pub notes: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SignatureDiffReportV0 {
-    pub schema_version: String,
-    pub before: SnapshotDiffEndpoint,
-    pub after: SnapshotDiffEndpoint,
-    pub comparison_status: SnapshotComparisonStatus,
-    pub delta_signature_signed: NullableSignatureIntVector,
-    pub metric_delta_status: BTreeMap<String, MetricDeltaStatus>,
-    pub worsened_axes: Vec<SignatureAxisChange>,
-    pub improved_axes: Vec<SignatureAxisChange>,
-    pub unchanged_axes: Vec<SignatureAxisChange>,
-    pub unmeasured_axes: Vec<UnmeasuredAxisDelta>,
-    pub evidence_diff: SnapshotEvidenceDiff,
-    pub attribution: SnapshotDiffAttribution,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirDocumentV0 {
-    pub schema_version: String,
-    pub architecture_id: String,
-    pub ids: AirIdPolicies,
-    pub revision: AirRevision,
-    pub feature: AirFeature,
-    pub artifacts: Vec<AirArtifact>,
-    pub evidence: Vec<AirEvidence>,
-    pub components: Vec<AirComponent>,
-    pub relations: Vec<AirRelation>,
-    pub policies: AirPolicies,
-    pub semantic_diagrams: Vec<AirSemanticDiagram>,
-    pub architecture_paths: Vec<AirArchitecturePath>,
-    pub homotopy_generators: Vec<AirHomotopyGenerator>,
-    pub nonfillability_witnesses: Vec<AirNonfillabilityWitness>,
-    pub signature: AirSignature,
-    pub coverage: AirCoverage,
-    pub claims: Vec<AirClaim>,
-    pub operation_trace: AirOperationTrace,
-    pub extension: AirExtension,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirIdPolicies {
-    pub component_id_policy: String,
-    pub relation_id_policy: String,
-    pub evidence_id_policy: String,
-    pub claim_id_policy: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirRevision {
-    pub before: Option<String>,
-    pub after: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirFeature {
-    pub feature_id: Option<String>,
-    pub title: Option<String>,
-    pub description: Option<String>,
-    pub source: String,
-    pub ai_session: Option<AirAiSession>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirAiSession {
-    pub provider: Option<String>,
-    pub model: Option<String>,
-    pub prompt_ref: Option<String>,
-    pub generated_patch: Option<bool>,
-    pub human_reviewed: Option<bool>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirArtifact {
-    pub artifact_id: String,
-    pub kind: String,
-    pub schema_version: Option<String>,
-    pub path: Option<String>,
-    pub content_hash: Option<String>,
-    pub produced_by: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirEvidence {
-    pub evidence_id: String,
-    pub kind: String,
-    pub artifact_ref: Option<String>,
-    pub path: Option<String>,
-    pub symbol: Option<String>,
-    pub line: Option<usize>,
-    pub rule_id: Option<String>,
-    pub confidence: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirComponent {
-    pub id: String,
-    pub kind: String,
-    pub lifecycle: String,
-    pub owner: Option<String>,
-    pub evidence_refs: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirRelation {
-    pub id: String,
-    pub layer: String,
-    #[serde(rename = "from")]
-    pub from_component: Option<String>,
-    #[serde(rename = "to")]
-    pub to_component: Option<String>,
-    pub kind: String,
-    pub lifecycle: String,
-    pub protected_by: Option<String>,
-    pub extraction_rule: Option<String>,
-    pub evidence_refs: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirPolicies {
-    pub laws: Vec<String>,
-    pub boundaries: Vec<String>,
-    pub allowed_edges: Vec<String>,
-    pub forbidden_edges: Vec<String>,
-    pub abstraction_rules: Vec<String>,
-    pub protection_rules: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirSemanticDiagram {
-    pub id: String,
-    pub lhs_path_ref: String,
-    pub rhs_path_ref: String,
-    pub equivalence: String,
-    pub filler_claim_ref: Option<String>,
-    pub nonfillability_witness_refs: Vec<String>,
-    pub observation_refs: Vec<String>,
-    pub lifecycle: String,
-    pub evidence_refs: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirArchitecturePath {
-    pub path_id: String,
-    pub source_state: String,
-    pub target_state: String,
-    pub steps: Vec<String>,
-    pub lifecycle: String,
-    pub evidence_refs: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirHomotopyGenerator {
-    pub generator_id: String,
-    pub kind: String,
-    pub diagram_refs: Vec<String>,
-    pub preserves_observation_claim_refs: Vec<String>,
-    pub evidence_refs: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirNonfillabilityWitness {
-    pub witness_id: String,
-    pub diagram_ref: String,
-    pub witness_kind: String,
-    pub claim_ref: String,
-    pub evidence_refs: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirSignature {
-    pub axes: Vec<AirSignatureAxis>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirSignatureAxis {
-    pub axis: String,
-    pub value: Option<i64>,
-    pub measured: bool,
-    pub measurement_boundary: String,
-    pub source: Option<String>,
-    pub reason: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirCoverage {
-    pub layers: Vec<AirCoverageLayer>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirCoverageLayer {
-    pub layer: String,
-    pub measurement_boundary: String,
-    pub universe_refs: Vec<String>,
-    pub measured_axes: Vec<String>,
-    pub unmeasured_axes: Vec<String>,
-    pub extraction_scope: Vec<String>,
-    pub exactness_assumptions: Vec<String>,
-    pub unsupported_constructs: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirClaim {
-    pub claim_id: String,
-    pub subject_ref: String,
-    pub predicate: String,
-    pub claim_level: String,
-    pub claim_classification: String,
-    pub measurement_boundary: String,
-    pub theorem_refs: Vec<String>,
-    pub evidence_refs: Vec<String>,
-    pub required_assumptions: Vec<String>,
-    pub coverage_assumptions: Vec<String>,
-    pub exactness_assumptions: Vec<String>,
-    pub missing_preconditions: Vec<String>,
-    pub non_conclusions: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirOperationTrace {
-    pub operations: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirExtension {
-    pub embedding_claim_ref: Option<String>,
-    pub feature_view_claim_ref: Option<String>,
-    pub interaction_claim_refs: Vec<String>,
-    pub split_claim_ref: Option<String>,
-    pub split_status: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct AirDocumentInput {
-    pub sig0_path: String,
-    pub validation_path: Option<String>,
-    pub diff_path: Option<String>,
-    pub pr_metadata_path: Option<String>,
-    pub law_policy_path: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirValidationReport {
-    pub schema_version: String,
-    pub input: AirValidationInput,
-    pub summary: AirValidationSummary,
-    pub checks: Vec<ValidationCheck>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub warnings: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirValidationInput {
-    pub schema_version: String,
-    pub path: String,
-    pub architecture_id: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AirValidationSummary {
-    pub result: String,
-    pub component_count: usize,
-    pub relation_count: usize,
-    pub evidence_count: usize,
-    pub claim_count: usize,
-    pub failed_check_count: usize,
-    pub warning_check_count: usize,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FeatureExtensionReportV0 {
-    pub schema_version: String,
-    pub input: FeatureReportInput,
-    pub architecture_id: String,
-    pub revision: AirRevision,
-    pub feature: AirFeature,
-    pub review_summary: FeatureReportReviewSummary,
-    pub architecture_summary: FeatureReportArchitectureSummary,
-    pub interpreted_extension: FeatureReportInterpretedExtension,
-    pub split_status: String,
-    pub preserved_invariants: Vec<FeatureReportInvariant>,
-    pub changed_invariants: Vec<FeatureReportInvariant>,
-    pub introduced_obstruction_witnesses: Vec<FeatureReportObstructionWitness>,
-    pub eliminated_obstruction_witnesses: Vec<FeatureReportObstructionWitness>,
-    pub complexity_transfer_candidates: Vec<String>,
-    pub semantic_path_summary: FeatureReportSemanticPathSummary,
-    pub theorem_package_refs: Vec<String>,
-    pub theorem_precondition_summary: TheoremPreconditionCheckSummary,
-    pub theorem_precondition_checks: Vec<TheoremPreconditionCheck>,
-    pub discharged_assumptions: Vec<String>,
-    pub undischarged_assumptions: Vec<String>,
-    pub coverage_gaps: Vec<FeatureReportCoverageGap>,
-    pub unsupported_constructs: Vec<String>,
-    pub repair_suggestions: Vec<String>,
-    pub empirical_annotations: Vec<String>,
-    pub non_conclusions: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FeatureReportInput {
-    pub schema_version: String,
-    pub path: String,
-    pub architecture_id: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FeatureReportReviewSummary {
-    pub split_status: String,
-    pub claim_classification: String,
-    pub top_witnesses: Vec<String>,
-    pub required_action: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FeatureReportArchitectureSummary {
-    pub component_count: usize,
-    pub relation_count: usize,
-    pub static_relation_count: usize,
-    pub runtime_relation_count: usize,
-    pub policy_relation_count: usize,
-    pub semantic_diagram_count: usize,
-    pub measured_axes: Vec<String>,
-    pub unmeasured_axes: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FeatureReportInterpretedExtension {
-    pub embedding_claim_ref: Option<String>,
-    pub feature_view_claim_ref: Option<String>,
-    pub interaction_claim_refs: Vec<String>,
-    pub split_claim_ref: Option<String>,
-    pub added_components: Vec<String>,
-    pub changed_components: Vec<String>,
-    pub added_relations: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FeatureReportInvariant {
-    pub invariant: String,
-    pub axis: String,
-    pub value: Option<i64>,
-    pub measurement_boundary: String,
-    pub evidence_refs: Vec<String>,
-    pub claim_refs: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FeatureReportObstructionWitness {
-    pub witness_id: String,
-    pub layer: String,
-    pub kind: String,
-    pub extension_role: Vec<String>,
-    pub extension_classification: Vec<String>,
-    pub components: Vec<String>,
-    pub edges: Vec<FeatureReportEdgeRef>,
-    pub paths: Vec<String>,
-    pub diagrams: Vec<String>,
-    pub nonfillability_witness_ref: Option<String>,
-    pub operation: Option<String>,
-    pub evidence: Vec<FeatureReportEvidenceRef>,
-    pub theorem_reference: Vec<String>,
-    pub claim_level: String,
-    pub claim_classification: String,
-    pub measurement_boundary: String,
-    pub non_conclusions: Vec<String>,
-    pub repair_candidates: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FeatureReportEdgeRef {
-    pub relation_id: String,
-    pub from: Option<String>,
-    pub to: Option<String>,
-    pub kind: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FeatureReportEvidenceRef {
-    pub evidence_ref: String,
-    pub kind: Option<String>,
-    pub artifact_ref: Option<String>,
-    pub path: Option<String>,
-    pub symbol: Option<String>,
-    pub rule_id: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FeatureReportSemanticPathSummary {
-    pub path_count: usize,
-    pub diagram_count: usize,
-    pub nonfillability_witness_count: usize,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FeatureReportCoverageGap {
-    pub layer: String,
-    pub measurement_boundary: String,
-    pub unmeasured_axes: Vec<String>,
-    pub unsupported_constructs: Vec<String>,
-    pub non_conclusions: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct TheoremPreconditionCheckReportV0 {
-    pub schema_version: String,
-    pub input: TheoremPreconditionCheckInput,
-    pub registry: TheoremPackageRegistryV0,
-    pub summary: TheoremPreconditionCheckSummary,
-    pub checks: Vec<TheoremPreconditionCheck>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct TheoremPreconditionCheckInput {
-    pub schema_version: String,
-    pub path: String,
-    pub architecture_id: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct TheoremPackageRegistryV0 {
-    pub schema_version: String,
-    pub scope: String,
-    pub packages: Vec<TheoremPackageMetadataV0>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct TheoremPackageMetadataV0 {
-    pub package_id: String,
-    pub lean_entrypoint: String,
-    pub theorem_refs: Vec<String>,
-    pub supported_subject_refs: Vec<String>,
-    pub supported_axes: Vec<String>,
-    pub claim_level: String,
-    pub claim_classification: String,
-    pub measurement_boundary: String,
-    pub required_assumptions: Vec<String>,
-    pub coverage_assumptions: Vec<String>,
-    pub exactness_assumptions: Vec<String>,
-    pub missing_preconditions: Vec<String>,
-    pub non_conclusions: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct TheoremPreconditionCheckSummary {
-    pub result: String,
-    pub checked_claim_count: usize,
-    pub applicable_claim_count: usize,
-    pub formal_proved_claim_count: usize,
-    pub measured_witness_count: usize,
-    pub blocked_claim_count: usize,
-    pub unmeasured_claim_count: usize,
-    pub unknown_theorem_ref_count: usize,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct TheoremPreconditionCheck {
-    pub claim_id: String,
-    pub subject_ref: String,
-    pub applicable_package_refs: Vec<String>,
-    pub theorem_refs: Vec<String>,
-    pub unknown_theorem_refs: Vec<String>,
-    pub claim_level: String,
-    pub input_claim_classification: String,
-    pub resolved_claim_classification: String,
-    pub measurement_boundary: String,
-    pub required_assumptions: Vec<String>,
-    pub coverage_assumptions: Vec<String>,
-    pub exactness_assumptions: Vec<String>,
-    pub missing_preconditions: Vec<String>,
-    pub non_conclusions: Vec<String>,
-    pub result: String,
-    pub reason: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SnapshotDiffEndpoint {
-    pub repository: SnapshotRepositoryRef,
-    pub revision: RepositoryRevisionRef,
-    pub validation_result: String,
-    pub extractor_version: String,
-    pub rule_set_version: String,
-    pub policy_id: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SnapshotComparisonStatus {
-    pub primary_diff_eligible: bool,
-    pub reasons: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SignatureAxisChange {
-    pub axis: String,
-    pub before: i64,
-    pub after: i64,
-    pub delta: i64,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UnmeasuredAxisDelta {
-    pub axis: String,
-    pub reason: String,
-    pub before_measured: bool,
-    pub after_measured: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SnapshotEvidenceDiff {
-    pub available: bool,
-    pub unavailable_reason: Option<String>,
-    pub component_delta: Option<ComponentSetDelta>,
-    pub edge_delta: Option<EdgeSetDelta>,
-    pub policy_violation_delta: Option<PolicyViolationSetDelta>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ComponentSetDelta {
-    pub before_count: usize,
-    pub after_count: usize,
-    pub delta: i64,
-    pub added: Vec<String>,
-    pub removed: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct EdgeSetDelta {
-    pub before_count: usize,
-    pub after_count: usize,
-    pub delta: i64,
-    pub added: Vec<Edge>,
-    pub removed: Vec<Edge>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct PolicyViolationSetDelta {
-    pub before_count: usize,
-    pub after_count: usize,
-    pub delta: i64,
-    pub added: Vec<PolicyViolation>,
-    pub removed: Vec<PolicyViolation>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SnapshotDiffAttribution {
-    pub method: String,
-    pub shared_worsened_axes: Vec<String>,
-    pub candidates: Vec<AttributionCandidate>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AttributionCandidate {
-    pub kind: String,
-    pub id: String,
-    pub confidence: f64,
-    pub confidence_level: String,
-    pub reasons: Vec<String>,
-    pub changed_components: Vec<String>,
-    pub matched_components: Vec<String>,
-    pub matched_edges: Vec<Edge>,
-    pub matched_policy_violations: Vec<PolicyViolation>,
-    pub affected_axes: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct NullableSignatureIntVector {
-    pub has_cycle: Option<i64>,
-    pub scc_max_size: Option<i64>,
-    pub max_depth: Option<i64>,
-    pub fanout_risk: Option<i64>,
-    pub boundary_violation_count: Option<i64>,
-    pub abstraction_violation_count: Option<i64>,
-    pub scc_excess_size: Option<i64>,
-    pub max_fanout: Option<i64>,
-    pub reachable_cone_size: Option<i64>,
-    pub weighted_scc_risk: Option<i64>,
-    pub projection_soundness_violation: Option<i64>,
-    pub lsp_violation_count: Option<i64>,
-    pub nilpotency_index: Option<i64>,
-    pub runtime_propagation: Option<i64>,
-    pub relation_complexity: Option<i64>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct MetricDeltaStatus {
-    pub comparable: bool,
-    pub reason: String,
-    pub before_measured: bool,
-    pub after_measured: bool,
-}
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 struct ImportEdge {
